Move and document MaybeFileBuf
